BUG: El càlcul de l'espai d'un DVD es feia malament, ja que la capacitat en bytes d'un DVD no es pot expressar en unsigned long int. S'ha passat d'unsigned long int a double. Ticket: #326. Tambe queda solucionat la resta de problemes del ticket, ja que esdevenia un overflow al unsigned long al calcular les diferents capacitats.

+++ starviewer/src/inputoutput/cacheimagedal.cpp
@@ -193,7 +193,7 @@ Status CacheImageDAL::countImageNumber( DicomMask imageMask , int &imageNumber )
    return state;
 }
 
-Status CacheImageDAL::imageSize (  DicomMask imageMask , unsigned long &size )
+Status CacheImageDAL::imageSize (  DicomMask imageMask , double &size )
 {
     int columns , rows , stateDatabase;
     char **reply = NULL , **error = NULL;
@@ -220,7 +220,7 @@ Status CacheImageDAL::imageSize (  DicomMask imageMask , unsigned long &size )
 
     if ( reply[1] != NULL )
     {
-        size = atol( reply [1] );
+        size = atof( reply [1] );
     }
     else size = 0;
